created a service function to validate errors for save table
